Add separation of interface and implementation
