+++ my_parser/CMakeLists.txt
@@ -3,4 +3,4 @@ project(my_parser)
 
 set(CMAKE_CXX_STANDARD 17)
 
-add_executable(my_parser main.cpp)
+add_executable(my_parser main.cpp parser.h parser.cpp)
