+++ my_parser/bad_test.txt
@@ -3,5 +3,5 @@
 2,=1+B2,=1   + C2,=1+D2,=1+E2,=1+F2,=1+G2,=1+H2,=1+I2,=1+J2,=1+A3
 3,=1+B3,=1+C3,=1+D3,=1+E3,=1+F3,=1+G3,=1+H3,=1+I3,=1+J3,=1+A4
 4,=1+B4,=1+C4,=1+D4,=1+E4,=1+F4,=1+G4,=1+H4,=1+I4,=1+J4,=1+A5
-5,=1+B5,=1+C5,=1+   D5   , =1+E5,=1+F5,=1+G5,=1+H5,=1+I5,=1+J5,=1+A6
+5,=1+B5,=1+C5,=1+   D5   , =1+E5,=1++F5,=1+G5,=1+H5,=1+I5,=1+J5,=1+A6
 6,=1+B6,=1+C6,=1+D6,=1+E6,=1+F6,=1+G6,=1+H6,=1+I6,=1/J6,1
