+++ my_parser/cmake-build-debug/CMakeFiles/Makefile2
@@ -81,12 +81,12 @@ clean: CMakeFiles/my_parser.dir/clean
 CMakeFiles/my_parser.dir/all:
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/my_parser.dir/build.make CMakeFiles/my_parser.dir/depend
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/my_parser.dir/build.make CMakeFiles/my_parser.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles --progress-num=1,2 "Built target my_parser"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles --progress-num=1,2,3 "Built target my_parser"
 .PHONY : CMakeFiles/my_parser.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/my_parser.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles 2
+	$(CMAKE_COMMAND) -E cmake_progress_start /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles 3
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/Makefile2 CMakeFiles/my_parser.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles 0
 .PHONY : CMakeFiles/my_parser.dir/rule
