+++ my_parser/cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,16 +1,4 @@
 /snap/clion/222/bin/cmake/linux/x64/bin/cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_MAKE_PROGRAM=/usr/bin/make -DCMAKE_C_COMPILER=/usr/bin/cc -DCMAKE_CXX_COMPILER=/usr/bin/g++-10 -G "CodeBlocks - Unix Makefiles" -S /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser -B /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug
--- The C compiler identification is GNU 11.3.0
--- The CXX compiler identification is GNU 10.4.0
--- Detecting C compiler ABI info
--- Detecting C compiler ABI info - done
--- Check for working C compiler: /usr/bin/cc - skipped
--- Detecting C compile features
--- Detecting C compile features - done
--- Detecting CXX compiler ABI info
--- Detecting CXX compiler ABI info - done
--- Check for working CXX compiler: /usr/bin/g++-10 - skipped
--- Detecting CXX compile features
--- Detecting CXX compile features - done
 -- Configuring done
 -- Generating done
 -- Build files have been written to: /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug
