+++ my_parser/cmake-build-debug/CMakeFiles/my_parser.dir/DependInfo.cmake
@@ -9,6 +9,7 @@ set(CMAKE_DEPENDS_LANGUAGES
 # The set of dependency files which are needed:
 set(CMAKE_DEPENDS_DEPENDENCY_FILES
   "/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/main.cpp" "CMakeFiles/my_parser.dir/main.cpp.o" "gcc" "CMakeFiles/my_parser.dir/main.cpp.o.d"
+  "/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/parser.cpp" "CMakeFiles/my_parser.dir/parser.cpp.o" "gcc" "CMakeFiles/my_parser.dir/parser.cpp.o.d"
   )
 
 # Targets to which this target links.
