+++ my_parser/cmake-build-debug/CMakeFiles/my_parser.dir/build.make
@@ -83,17 +83,33 @@ CMakeFiles/my_parser.dir/main.cpp.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/my_parser.dir/main.cpp.s"
 	/usr/bin/g++-10 $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/main.cpp -o CMakeFiles/my_parser.dir/main.cpp.s
 
+CMakeFiles/my_parser.dir/parser.cpp.o: CMakeFiles/my_parser.dir/flags.make
+CMakeFiles/my_parser.dir/parser.cpp.o: /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/parser.cpp
+CMakeFiles/my_parser.dir/parser.cpp.o: CMakeFiles/my_parser.dir/compiler_depend.ts
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Building CXX object CMakeFiles/my_parser.dir/parser.cpp.o"
+	/usr/bin/g++-10 $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -MD -MT CMakeFiles/my_parser.dir/parser.cpp.o -MF CMakeFiles/my_parser.dir/parser.cpp.o.d -o CMakeFiles/my_parser.dir/parser.cpp.o -c /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/parser.cpp
+
+CMakeFiles/my_parser.dir/parser.cpp.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/my_parser.dir/parser.cpp.i"
+	/usr/bin/g++-10 $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/parser.cpp > CMakeFiles/my_parser.dir/parser.cpp.i
+
+CMakeFiles/my_parser.dir/parser.cpp.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/my_parser.dir/parser.cpp.s"
+	/usr/bin/g++-10 $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/parser.cpp -o CMakeFiles/my_parser.dir/parser.cpp.s
+
 # Object files for target my_parser
 my_parser_OBJECTS = \
-"CMakeFiles/my_parser.dir/main.cpp.o"
+"CMakeFiles/my_parser.dir/main.cpp.o" \
+"CMakeFiles/my_parser.dir/parser.cpp.o"
 
 # External object files for target my_parser
 my_parser_EXTERNAL_OBJECTS =
 
 my_parser: CMakeFiles/my_parser.dir/main.cpp.o
+my_parser: CMakeFiles/my_parser.dir/parser.cpp.o
 my_parser: CMakeFiles/my_parser.dir/build.make
 my_parser: CMakeFiles/my_parser.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Linking CXX executable my_parser"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/aleksandr/data/course2/C++/Yadro_test_task/CSVParser/my_parser/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Linking CXX executable my_parser"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/my_parser.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
