+++ my_parser/cmake-build-debug/CMakeFiles/my_parser.dir/cmake_clean.cmake
@@ -1,6 +1,8 @@
 file(REMOVE_RECURSE
   "CMakeFiles/my_parser.dir/main.cpp.o"
   "CMakeFiles/my_parser.dir/main.cpp.o.d"
+  "CMakeFiles/my_parser.dir/parser.cpp.o"
+  "CMakeFiles/my_parser.dir/parser.cpp.o.d"
   "my_parser"
   "my_parser.pdb"
 )
