+++ my_parser/cmake-build-debug/CMakeFiles/my_parser.dir/link.txt
@@ -1,1 +1,1 @@
-/usr/bin/g++-10 -g CMakeFiles/my_parser.dir/main.cpp.o -o my_parser 
+/usr/bin/g++-10 -g CMakeFiles/my_parser.dir/main.cpp.o CMakeFiles/my_parser.dir/parser.cpp.o -o my_parser
