+++ my_parser/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Feb 20 23:47 MSK
+Start testing: Feb 21 00:19 MSK
 ----------------------------------------------------------
-End testing: Feb 20 23:47 MSK
+End testing: Feb 21 00:19 MSK
